include automaticlly lougout & make sure if the session has bee setn

+++ removec.php
@@ -1,4 +1,9 @@
 <?php
+      if (!isset($_SESSION['signin'])) {
+        header("location: index.php");
+        exit();
+      }
+      include './includes/time-logout.php';
       include './includes/db_conn.php';
       $id = $_GET['id'];
       $remove = "DELETE FROM course WHERE id = $id";
